"1.3-as received.rb" May 6, 2006, 5:33 PM
Again, no idea what's up with the order here and I'm kind of guessing.

I appear to have sent him some questions, in a file called
"1.3questions" with a creation date of May 3, 7:49 AM (?????) Contents below:

------

Line 39:
Dir.mkdir($fiendDir) unless File.exists?($fiendDir)

Since we're not covering for a nonexistant fiend.txt, I was wondering about this. My guess was so that it's less work for the user to make a fiend.txt (user doesn't have to do a mkdir themselves), but do we really want to be leaving marks behind if we're going to fail out of the program anyhow?

Lines 78-80:
       if ((line =~ /^(\^[\S]+) (.*)$/) != nil)
               tag = $1.downcase
               entry = $2
       end
Could you give a real quick explanation of this? I can see what the result is, but I haven't seen this construction before.

Lines 83-84:
       when /^#/, "", /^;/, /\/\//
               fiendTwigs << line + "\n"
What's this?
------