An experimental rewrite of the save script in Ruby
So here's the twin problems with automating Mac apps:

- Doing any normal-ass programming shit (like munging a string) is pure torture in Applescript.
- Shelling out to `osascript` performs like crap. Like, it's shockingly bad.

So, this script works and is legible, but it's so slow that I can't consider
using it. While the original Applescript version is really a disaster to look
at, but it's nice and fast and works fine as long as you don't squint at it. I
haven't really changed it in years, so I probably could have let it be.