Javascript version of save script
So then I remembered that you can use Javascript in place of Applescript.

Now, last time I used this, I was really impressed with its capabilities, and
really UNIMPRESSED with its lack of documentation and shoddy support. But it
really does match my needs pretty exactly (a fast connection to the scripting
bridge, while also being a real-ish programming language that doesn't require
mind-reading to write), so I gave it another go.

Does anyone know if they did any upkeep on this in Mojave? I feel like it's in
somewhat better shape, or at least I seem to remember trying a few of the things
in here before and having it fail mysteriously. Anyway, other than some
slapstick with Javascript's built-in Date objects, and a bit of a leap of faith
re: how to access an application's contained objects (`bb.textDocuments[i]`),
this actually went incredibly smoothly.

Remember that these text representations have to be saved as a "script" with
Script Editor before they'll run well; that way they get saved in a compiled
form and get to go fast.