FMP dump file append script
So back in the day, I used the Quicksilver launcher, which eventually became
sort of a legendary monument of weird abandonware standing out in the desert
with an ironic epigram chiseled into it, or... something like that. Anyway, it
had a cool "append to text file" action, and I had a hotkey that would bring it
up with my FMP dump file already selected and the cursor ready to go.

Once QS started rotting in earnest I switched to LaunchBar, which is less
all-singing-all-dancing but which is at least actively maintained (and does a
couple of things in a nicer way than QS used to). LB also has an append action,
but for some fucking reason it's impossible to script it into opening, so
instead of "hotkey -> type a line and hit enter", it was "hotkey -> cmd-shift-A
-> type a line and hit enter". Which eventually got too annoying, because my
fingers still remembered how it was SUPPOSED to work, so I decided to try and
roll my own. (It's possible that Alfred or one of the other launchers has an
append action that would fit the bill, but this seemed more interesting.)

I never bothered to try this type of thing with AppleScript because I knew THAT
was going to be a nightmare, but this seemed like a good chance to find out
whether Yosemite's new "Javascript for Automation" scripting was any good.
Verdict: kinda-sorta. I DO love that it lets you create ObjC values and use the
Cocoa APIs whenever you need to do something that AppleScript would be really
shit at. And god it's nice to have a normal-ass language to work with (for some
values of normal and some values of ass). But it's really underdocumented (it's
especially hard to figure out how to navigate through a scriptable application's
properties and actions), and I can't tell how much I should trust Apple to
improve it and keep it working. It might have gotten a little bit better in
Mojave; I can't quite be sure yet.

Anyway, this script works great and I've been using it for years now. There's a
small issue where sometimes it'll trigger twice, so the dialog will pop up again
as soon as I hit enter; I'm not sure if that's just a keyboard sensitivity issue
with FastScripts or something else, but either way I can just hit esc to dismiss
the second one.