Make decompilation better
OK, I initially tried ~being clever~ with this, by using `rule`s to teach Rake
how to compile AND decompile scripts and shortcuts. I figured it'd just treat
the rule as a subroutine, basically. But instead, it uses the rule to synthesize
a new task the first time something that matches the rule gets mentioned as a
dependency. And because the files go round-trip, the decompilation rule was
referencing compiled files as dependencies and vice versa, which would create a
new task, which would create a new task... So if I was using good hygiene for
dealing with file extensions, it'd recognize the issue and bail with a circular
dependency error, and if I was being lazy with extensions, it'd keep tacking on
extra .applescript.applescript.applescript.js bits and then bail with a
recursion limit error.

Anyway, be stupid instead of being clever, you'll have a better life. But I
eventually won and now it's fairly easy to make updates to shortcuts after
editing them in ios.