iOS FMP: Make direct-append shortcut handle multiple lines better
If you entered something with linebreaks into the text field, and at least one
of those lines started with a ^caret-tag, I think the first version of this
would throw away everything except that first line. I think. Anyway, now we
split the text and handle each line separately, so you can append a whole bunch
of crap at once to a bunch of different files if you're feelin' it.