iOS FMP errata: chunked refresh
iA Writer's URL scheme documentation claims that there's a 4000 character limit
(before URL encoding) for url parameter values, and my fiend.txt was more like
9000 characters (I've cleaned it up since then), and during my initial testing I
managed to crash iA Writer and was worried that it was the character limit that
was doing it. So I copied the refresh script and made a version that would
rewrite fiend.txt in chunks, to make sure we stayed under the limit.

But really it turns out that the limit is more of a suggestion, or maybe it came
from a previous version of iOS's URL handling and Apple bumped up the limits
because they knew that the big relaunch of the Shortcuts app was going to mean a
lot more people doing weird stuff with URL schemes. I don't know! I also don't
know what crashed my app that one time (are there different memory limits when
you run a shortcut from the edit view, rather than the launcher view? Who
knows!). What I DO know is that there's no good reason to keep this extra
complexity, so I went back to the non-chunked version of the script. This has
been hanging around in my shortcuts launcher ever since, and I wanted to back it
up somewhere for posterity before I deleted it. (I feel like my chunked writes
solution was mildly clever... but, like, clever in a bad way.)